Module asn1_cereal::ber::enc

Encoding rules to use when encoding or decoding ASN.1.

The defaults are to use the DER encoding during serialization, and to use the BER encoding during deserliazation (all valid BER is accepted, including DER).

Structs

BER
Basic Encoding Rules define the most basic rules that can be used to encode an ASN.1 tag. These are the default encoding rules used when decoding ASN1, as all variants are valid BER.

BERAlt
This set of rules creates a valid BER stream, but will use implicit tags where possible, and indefinite length encoding for all constructed elements.

DER
Distinguished Encoding Rules are a subset of BER, and provide a deterministic, shortest form of encoding. These are the default encoding rules used when encoding ASN.1.

Enums

LenEnc
An enum which determines how lengths are encoded.

TagEnc
An enum which determines how tags are encoded.

Traits

BerEncRules
A trait to define encoding rules to use while encoding ASN.1.