Struct asn1_cereal::ber::enc::BER

pub struct BER;

Basic Encoding Rules define the most basic rules that can be used to encode an ASN.1 tag. These are the default encoding rules used when decoding ASN1, as all variants are valid BER.

When encoding, explicit tags and definite length encoding will be used for all elements.