Trait asn1_cereal::ber::enc::BerEncRules

pub trait BerEncRules: Copy {
    fn tag_rules() -> TagEnc;
    fn len_rules() -> LenEnc;
    fn shortest_form() -> bool;
}

A trait to define encoding rules to use while encoding ASN.1.

Required Methods

fn tag_rules() -> TagEnc

Get the encoding rules for tags.

fn len_rules() -> LenEnc

Get the encoding rules for lengths.

fn shortest_form() -> bool

Returns true when primitive elements should be encoded using the shortest form. If this is true when decoding, it should be an error to not use the shortest form.