Module asn1_cereal::ber::serial

Traits for serializing and deserializing rust types from/to ASN.1.

The traits in this module can be used to implement serialization and deserialization for a Rust type from/to ASN.1. For built-in types, this module provides an implementation already. For more complex types (like structs and enumms), macros are provided to generate an implementation for them.

Usage

The below example defines a Rust struct, and uses a macro to generate the code required to serialize it to an ASN.1 sequence, or to deserialize an ASN.1 sequence into an instance of the struct.

The example then uses the BerSerialize::serialize function to encode the struct to a DER-encoded byte stream.

#[macro_use] extern crate asn1_cereal; fn main() {
  struct ShortSequence {
    z: u64,
    y: u32,
  }

  ber_sequence!(
    ShortSequence,
    "SHORT_SEQUENCE",
    z;
    y;
  );

  use asn1_cereal::BerSerialize;

  let data = ShortSequence{ z: 1, y: 2 };
  let mut bytes: Vec<u8> = Vec::new();
  BerSerialize::serialize(&data, &mut bytes).unwrap();
}

Modules

alias
Macros to generate the implementation of the serialization traits for Rust newtypes (tuple structs). This can be used to represent ASN.1 type assignments, or wrap Rust types, to provide distinct types. ```

bool
Implementation of the serialization traits for Rust booleans.

choice
Macros to generate the implementation of the serialization traits for Rust enums, as ASN.1 choice.

int
Implementation of the serialization traits for Rust integers.

prim
A collection of primitive wrappers for ASN.1, where a direct Rust equivalent may not exist.

seq
Macros to generate the implementation of the serialization traits for Rust structs, as an ASN.1 SEQUENCE. You can technically use this for a SET, but it still assumes the elements are in order.

seq_of
Macros to generate the implementation of the serialization traits for Rust iterators, as ASN.1 sequence of.

str
Implementation of the serialization traits for String.

traits
The base traits that are used for serializing and deserializing rust types.