Trait asn1_cereal::ber::serial::traits::BerDeserialize

pub trait BerDeserialize: Asn1Info + Sized {
    fn deserialize<I: Iterator<Item=Result<u8>>>(reader: &mut I) -> Result<Self, DecodeError> { ... }
    fn deserialize_enc<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I) -> Result<Self, DecodeError> { ... }
    fn deserialize_with_tag<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag, len: Len) -> Result<Self, DecodeError> { ... }
    fn _deserialize_with_tag<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag, len: Len) -> Option<Result<Self, DecodeError>> { ... }
    fn deserialize_value<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, len: Len) -> Result<Self, DecodeError> { ... }
}

Provides the methods required to deserialize this Rust type from an ASN.1 stream.

When implementing this for a simple primitive type, implementing deserialize_value should be all that's required.

Provided Methods

fn deserialize<I: Iterator<Item=Result<u8>>>(reader: &mut I) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, accepting any valid BER.

fn deserialize_enc<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, using a specific set of encoding rules.

fn deserialize_with_tag<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag, len: Len) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, using a specific set of encoding rules, and also providing the decoded tag and length.

This function assumes the next bytes to decode are the BER length of this element.

fn _deserialize_with_tag<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag, len: Len) -> Option<Result<Self, DecodeError>>

fn deserialize_value<E: BerEncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, len: Len) -> Result<Self, DecodeError>

Deserialize an ASN.1 value from a BER stream, after having the tag and length decoded.

The data length must be explicitly passed to this function. For primitive types, an error will be returned if this length is Indefinite.