Struct asn1_cereal::enc::DER

pub struct DER;

Distinguished Encoding Rules are a subset of BER, and provide a deterministic, shortest form of encoding. These are the default encoding rules used when encoding ASN.1.

Note: When decoding elements, invalid DER will cause an error. If this is an issue, the BER encoding rules should be used