Module asn1_cereal::serial::seq

The macros in this module can implement the serialisation traits for Sequences and Sets.

#[macro_use] extern crate asn1_cereal; fn main() {
  struct SomeSequence {
    a: u64,
    b: u32,
    c: String,
  }

  asn1_sequence_info!(SomeSequence, "SOME_SEQUENCE");
  asn1_sequence_serialize!(SomeSequence, a, b, c);
  asn1_sequence_deserialize!(SomeSequence, a, b, c);
}

Note that these macros can also be used for a SET, but not a SEQUENECE OF, or SET OF. (These are more appropriate as Vec or Set).