Trait asn1_cereal::serial::Asn1Deserialize

pub trait Asn1Deserialize: Asn1Info + Sized {
    fn deserialize_bytes<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, len: Option<LenNum>) -> Result<Self, DecodeError>;

    fn deserialize<I: Iterator<Item=Result<u8>>>(reader: &mut I) -> Result<Self, DecodeError> { ... }
    fn deserialize_enc<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I) -> Result<Self, DecodeError> { ... }
    fn deserialize_enc_tag<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag) -> Result<Self, DecodeError> { ... }
}

Provides the methods required to deserialize this Rust type from an ASN.1 stream.

When implementing this for a simple primitive type, implementing deserialize_bytes should be all that's required.

Required Methods

fn deserialize_bytes<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, len: Option<LenNum>) -> Result<Self, DecodeError>

Deserialise ASN.1 data without a tag into a value.

The data length must be explicitly passed to this function. For primitive types, an error will be returned if this length is Indefinite.

Provided Methods

fn deserialize<I: Iterator<Item=Result<u8>>>(reader: &mut I) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, accepting any valid BER.

fn deserialize_enc<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, using a specific set of encoding rules.

fn deserialize_enc_tag<E: Asn1EncRules, I: Iterator<Item=Result<u8>>>(e: E, reader: &mut I, tag: Tag) -> Result<Self, DecodeError>

Deserialize ASN.1 data into a Rust value, using a specific set of encoding rules, and also providing the decoded tag.

This function assumes the next bytes to decode are the BER length of this element.