Enum asn1_cereal::tag::Len

pub enum Len {
    Def(u64),
    Indef,
}

An enum representing the length of an ASN.1 element.

Variants
Def(u64)

A Definite length element.

Indef

An Indefinite length element, not known before decoding.

Methods

impl Len

fn read_indef_end<I: Iterator<Item=Result<u8>>>(bytes: &mut I) -> Result<(), DecodeError>

Given an iterator, read an indefinite length terminator.

fn write_indef_end<W: Write>(writer: &mut W) -> Result<(), EncodeError>

Write an indefinite length terminator to the given writer.

fn read_len<I: Iterator<Item=Result<u8>>>(bytes: &mut I) -> Result<Self, DecodeError>

Given an iterator over a byte stream, read and return the ASN.1 element length.

fn write_len<W: Write>(self, writer: &mut W) -> Result<(), EncodeError>

Write this ASN.1 length to the given writer.

fn as_num(self) -> Option<LenNum>

A short-hand function to quickly get an Option.