Crate asn1_cereal

Modules

ber

byte

err

serial

tag

Macros

asn1_info!
This macro is a compact way of defining asn1_info for a type.

asn1_newtype_deserialize!
This macro will generate an Asn1Deserialize implementation for an anonymous struct (a newtype).

asn1_newtype_serialize!
This macro will generate an Asn1Serialize implementation for an anonymous struct (a newtype).

asn1_sequence_deserialize!
This macro will generate an Asn1Deserialize implementation for a struct that represents an ASN.1 structure.

asn1_sequence_info!
This macro is a compact way of defining the Asn1Info trait implementation for a struct that represents an ASN.1 structure. If a custom class or tag number is required, asn1_info! should be used instead.

asn1_sequence_serialize!
This macro will generate an Asn1Serialize implementation for a struct that represents an ASN.1 structure.