Crate asn1_cereal

asn1-cereal

A collection of encoders and decoders for BER, DER and ASN.1.

The grains of this library are a collection of traits and macros, that allow serialization and deserialization of rust types to and from ASN.1.

The Asn1Info, BerSerialize and BerDeserialize traits are what most users will want to use.

Reexports

pub use info::Asn1Info;

pub use ber::serial::traits::{BerSerialize, BerDeserialize};

Modules

ber
Tools that can be used to parse BER streams.

byte
Constructs for reading and writing bytes used by this crate.

err
Encoding and Decoding errors that this crate can produce.

info

tag
Encoding and decoding of tag and length bytes for BER.

Macros

asn1_choice!

asn1_choice_deserialize!

asn1_choice_serialize!

asn1_info!
This macro defines the Asn1Info trait for a rust type.

asn1_newtype!
This macro is a compact way of defining both of the Asn1 serialization traits - BerSerialize and BerDeserialize - for a rust newtype, that represents an ASN.1 type definition.

asn1_newtype_deserialize!
This macro defines the BerSerialize trait for a rust newtype.

asn1_newtype_serialize!
This macro defines the BerSerialize trait for a rust newtype.

asn1_sequence!
This macro is a compact way of defining all three of the Asn1 traits - Asn1Info, BerSerialize and BerDeserialize - for a rust struct, that represents an ASN.1 sequence.

asn1_sequence_deserialize!
This macro defines the BerDeserialize trait for a rust struct. The code generated will deserialize the specified fields in the order that they are given.

asn1_sequence_info!
This macro defines the Asn1Info trait for a rust struct. This allows the other traits to get information about this type. If you need to provide a custom class or tag, consider using the asn1_info! macro.

asn1_sequence_of!

asn1_sequence_of_deserialize!

asn1_sequence_of_serialize!

asn1_sequence_serialize!
This macro defines the BerSerialize trait for a rust struct. The code generated will serialize the specified fields in the order that they are given.